requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-debug-toolbar==4.3.0
black==24.4.2
flake8==7.0.0
flake8-quotes==3.4.0
flake8-variables-names==0.0.6
pep8-naming==0.13.3
Pillow==10.3.0
python-dotenv==1.0.1
